file names and env variable for failover
