+++ rmf_fleet_adapter/src/robot_state_aggregator/main.cpp
@@ -1,178 +1,102 @@
-/*
- * Copyright (C) 2019 Open Source Robotics Foundation
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
-*/
+// Copyright 2016 Open Source Robotics Foundation, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
 
-#include <rclcpp/rclcpp.hpp>
-#include "rclcpp_components/register_node_macro.hpp"
+#include <memory>
+#include <string>
+#include <vector>
+#include <iostream>
 
-#include <rmf_fleet_msgs/msg/robot_state.hpp>
-#include <rmf_fleet_msgs/msg/fleet_state.hpp>
+#include "class_loader/class_loader.hpp"
+#include "rclcpp/rclcpp.hpp"
+#include "rclcpp_components/node_factory.hpp"
 
-#include <rmf_fleet_adapter/StandardNames.hpp>
+#define LINKTIME_COMPOSITION_LOGGER_NAME "linktime_composition"
 
-#ifdef FAILOVER_MODE
-#include "stubborn_buddies_msgs/msg/status.hpp"
-#endif
-
-using RobotState = rmf_fleet_msgs::msg::RobotState;
-using FleetState = rmf_fleet_msgs::msg::FleetState;
-
-
-#ifdef FAILOVER_MODE
-constexpr char DEFAULT_ACTIVE_STATUS_NAME[] = "status";
-#endif
-
-class RobotStateAggregator : public rclcpp::Node
+int main(int argc, char* argv[])
 {
-public:
+  // Force flush of the stdout buffer.
+  setvbuf(stdout, NULL, _IONBF, BUFSIZ);
 
-  explicit RobotStateAggregator(const rclcpp::NodeOptions& options)
-  : rclcpp::Node("robot_state_aggregator", options)
+  rclcpp::init(argc, argv);
+  rclcpp::Logger logger = rclcpp::get_logger(LINKTIME_COMPOSITION_LOGGER_NAME);
+  rclcpp::executors::SingleThreadedExecutor exec;
+  rclcpp::NodeOptions options;
+  std::vector<class_loader::ClassLoader*> loaders;
+  std::vector<rclcpp_components::NodeInstanceWrapper> node_wrappers;
+
+  std::vector<std::string> libraries = 
   {
-    RCLCPP_WARN(get_logger(), "RobotStateAggregator called");
-    const auto default_qos = rclcpp::SystemDefaultsQoS();
-    const auto sensor_qos = rclcpp::SensorDataQoS();
+    // all classes from libraries linked by the linker (rather then dlopen)
+    // are registered under the library_path ""
+    "",
+  };
 
-#ifdef FAILOVER_MODE
-    _active_node = this->declare_parameter("active_node", true);
-    _failover_mode = this->declare_parameter("failover_mode", false);
-    _buddy_subns = this->declare_parameter("buddy_subns", "yang");
-    _namespace = this->declare_parameter("namespace", "node_ns");
+  //Easy way to catch the failover parameter
+  auto dummy_node = std::make_shared<rclcpp::Node>("dumy_node");
+  bool failover_mode = dummy_node->declare_parameter("failover_mode", false);
+  dummy_node.reset();
 
-    _active_status_topic = "/" + _namespace +
-      "/" + _buddy_subns + "/"
-      + DEFAULT_ACTIVE_STATUS_NAME;
-#endif
-
-    _fleet_state_pub = create_publisher<FleetState>(
-      rmf_fleet_adapter::FleetStateTopicName, default_qos);
-
-#ifdef FAILOVER_MODE
-    if (_active_node)
-    {
-#endif
-      _robot_state_sub = create_subscription<RobotState>(
-        "/robot_state", sensor_qos,
-        [&](RobotState::UniquePtr msg)
-        {
-          _robot_state_update(std::move(msg));
-        });
-#ifdef FAILOVER_MODE
-    }
-    else
-    {
-      _inactive_state_sub = create_subscription
-        <stubborn_buddies_msgs::msg::Status>(
-        _active_status_topic,
-        10,
-        [this](const typename stubborn_buddies_msgs::msg::
-        Status::SharedPtr msg) -> void
-        {
-          const auto sensor_qos = rclcpp::SensorDataQoS();
-          RCLCPP_WARN(get_logger(), "Watchdog rised at %s, "
-          "self activation triggered",
-          _active_status_topic.c_str(),
-          msg->stamp.sec);
-          this->set_parameter(rclcpp::Parameter("active_node", true));
-          _active_node = true;
-
-          _robot_state_sub = create_subscription<RobotState>(
-            "/robot_state", sensor_qos,
-            [&](RobotState::UniquePtr msg)
-            {
-              _robot_state_update(std::move(msg));
-            });
-        });
-    }
-#endif
-
-    const auto prefix = this->declare_parameter("robot_prefix", "");
-    const auto fleet_name = this->declare_parameter("fleet_name", "");
-
-    if (fleet_name.empty())
-    {
-      RCLCPP_FATAL(
-        this->get_logger(),
-        "Missing required parameter: [fleet_name]");
-      exit(1);
-    }
-
-    this->_prefix = std::move(prefix);
-    this->_fleet_name = std::move(fleet_name);
+#ifndef FAILOVER_MODE
+  if (failover_mode)
+  {
+    RCLCPP_ERROR(logger, "robot_state_aggregator was compiled without failover"
+                 " support. Make sure you have the required libraries and the"
+                 " environment variable $RMF_ENABLE_FAILOVER is set during"
+                 " compilation.");
+    return 1;
   }
-
-private:
-  bool _active_node;
-
-  std::string _prefix;
-  std::string _fleet_name;
-  bool _failover_mode;
-#ifdef FAILOVER_MODE
-  std::string _active_status_topic;
-  std::string _buddy_subns;
-  std::string _namespace;
 #endif
 
-  std::unordered_map<std::string, std::unique_ptr<RobotState>> _latest_states;
-
-  rclcpp::Publisher<FleetState>::SharedPtr _fleet_state_pub;
-  rclcpp::Subscription<RobotState>::SharedPtr _robot_state_sub;
-
-#ifdef FAILOVER_MODE
-  rclcpp::Subscription<stubborn_buddies_msgs::msg::Status>::SharedPtr
-    _inactive_state_sub;
-#endif
-
-  void _robot_state_update(RobotState::UniquePtr msg)
+  for (auto library : libraries)
   {
-    const std::string& name = msg->name;
-    if (name.size() < _prefix.size())
-      return;
+    RCLCPP_INFO(logger, "Library");
 
-    if (name.substr(0, _prefix.size()) != _prefix)
-      return;
+    auto loader = new class_loader::ClassLoader(library);
+    auto classes =
+      loader->getAvailableClasses<rclcpp_components::NodeFactory>();
 
-    const auto insertion = _latest_states.insert(std::make_pair(name, nullptr));
-    const auto it = insertion.first;
-    bool updated = false;
-    if (insertion.second)
+    for (auto clazz : classes)
     {
-      it->second = std::move(msg);
-      updated = true;
-    }
-    else
-    {
-      if (rclcpp::Time(it->second->location.t) < rclcpp::Time(msg->location.t) )
+      RCLCPP_INFO(logger, "before if %s", clazz.c_str());
+      if (failover_mode || 
+         ((clazz.compare("rclcpp_components::NodeFactoryTemplate"
+         "<lifecycle_heartbeat::LifecycleHeartbeat>") != 0) &&
+         ((clazz.compare("rclcpp_components::NodeFactoryTemplate"
+         "<lifecycle_watchdog::LifecycleWatchdog>") != 0))))
       {
-        it->second = std::move(msg);
-        updated = true;
+        RCLCPP_INFO(logger, "Instantiate class %s", clazz.c_str());
+        auto node_factory =
+          loader->createInstance<rclcpp_components::NodeFactory>(clazz);
+        auto wrapper = node_factory->create_node_instance(options);
+        auto node = wrapper.get_node_base_interface();
+
+        node_wrappers.push_back(wrapper);
+        exec.add_node(node);
       }
-    }
 
-    if (updated)
-    {
-      FleetState fleet;
-      fleet.name = _fleet_name;
-      for (const auto& robot_state : _latest_states)
-        fleet.robots.emplace_back(*robot_state.second);
-
-      _fleet_state_pub->publish(fleet);
     }
+    loaders.push_back(loader);
   }
+  exec.spin();
 
-};
+  for (auto wrapper : node_wrappers)
+  {
+    exec.remove_node(wrapper.get_node_base_interface());
+  }
+  node_wrappers.clear();
 
-RCLCPP_COMPONENTS_REGISTER_NODE(RobotStateAggregator)
+  rclcpp::shutdown();
+
+  return 0;
+}
